+++ src/main/java/com/maxmind/geoip2/model/CityIspOrgResponse.java
@@ -10,7 +10,7 @@ package com.maxmind.geoip2.model;
  * @see <a href="http://dev.maxmind.com/geoip/geoip2/web-services">GeoIP2 Web
  *      Services</a>
  */
-final public class CityIspOrg extends AbstractCity {
-    public CityIspOrg() {
+final public class CityIspOrgResponse extends AbstractCityResponse {
+    public CityIspOrgResponse() {
     }
 }