+++ src/main/java/com/maxmind/geoip2/model/OmniResponse.java
@@ -10,7 +10,7 @@ package com.maxmind.geoip2.model;
  * @see <a href="http://dev.maxmind.com/geoip/geoip2/web-services">GeoIP2 Web
  *      Services</a>
  */
-final public class Omni extends AbstractCity {
-    public Omni() {
+final public class OmniResponse extends AbstractCityResponse {
+    public OmniResponse() {
     }
 }