+++ src/main/java/com/maxmind/geoip2/record/City.java
@@ -7,11 +7,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * 
  * This record is returned by all the end points except the Country end point.
  */
-final public class CityRecord extends AbstractNamedRecord {
+final public class City extends AbstractNamedRecord {
     @JsonProperty
     private Integer confidence;
 
-    public CityRecord() {
+    public City() {
         super();
     }