+++ src/main/java/com/maxmind/geoip2/record/RepresentedCountry.java
@@ -11,11 +11,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * 
  * This record is returned by all the end points.
  */
-final public class RepresentedCountryRecord extends CountryRecord {
+final public class RepresentedCountry extends Country {
     @JsonProperty
     private String type;
 
-    public RepresentedCountryRecord() {
+    public RepresentedCountry() {
     }
 
     /**