+++ src/test/java/com/maxmind/geoip2/DatabaseReaderTest.java
@@ -14,9 +14,9 @@ import org.junit.rules.ExpectedException;
 
 import com.maxmind.geoip2.exception.AddressNotFoundException;
 import com.maxmind.geoip2.exception.GeoIp2Exception;
-import com.maxmind.geoip2.model.City;
-import com.maxmind.geoip2.model.CityIspOrg;
-import com.maxmind.geoip2.model.Omni;
+import com.maxmind.geoip2.model.CityResponse;
+import com.maxmind.geoip2.model.CityIspOrgResponse;
+import com.maxmind.geoip2.model.OmniResponse;
 
 @SuppressWarnings("static-method")
 public class DatabaseReaderTest {
@@ -28,7 +28,7 @@ public class DatabaseReaderTest {
     public void testDefaultLanguage() throws IOException, GeoIp2Exception {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"));
-        City city = reader.city(InetAddress.getByName("81.2.69.160"));
+        CityResponse city = reader.city(InetAddress.getByName("81.2.69.160"));
         assertEquals("London", city.getCity().getName());
         reader.close();
     }
@@ -38,7 +38,7 @@ public class DatabaseReaderTest {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"), Arrays.asList("xx", "ru",
                 "pt-BR", "es", "en"));
-        Omni city = reader.omni(InetAddress.getByName("81.2.69.160"));
+        OmniResponse city = reader.omni(InetAddress.getByName("81.2.69.160"));
         assertEquals("Лондон", city.getCity().getName());
         reader.close();
 
@@ -48,7 +48,7 @@ public class DatabaseReaderTest {
     public void hasIpAddress() throws IOException, GeoIp2Exception {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"));
-        CityIspOrg cio = reader
+        CityIspOrgResponse cio = reader
                 .cityIspOrg(InetAddress.getByName("81.2.69.160"));
         assertEquals("81.2.69.160", cio.getTraits().getIpAddress());
         reader.close();
@@ -63,7 +63,7 @@ public class DatabaseReaderTest {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"));
         @SuppressWarnings("unused")
-        City city = reader.city(InetAddress.getByName("10.10.10.10"));
+        CityResponse city = reader.city(InetAddress.getByName("10.10.10.10"));
         reader.close();
     }